fix: do not allow 0 tokens or messages

+++ tests/chat.h
@@ -1,3 +1,4 @@
+#include <sys/_stdint.h>
 #ifndef chatGPTuino_h
 #define chatGPTuino_h
 
@@ -18,8 +19,6 @@ namespace ChatGPTuino {
 #define MAX_TOKENS 2000  // Used for sizing JSON reponse
 #define MIN_MESSAGES 5
 #define CHARS_PER_TOKEN 6
-// #define API_KEY_SIZE 100
-// #define MODEL_NAME_SIZE 50
 
 // These constants are used for calculating the size of Dynamic JSON Array
 // TODO Consider using the JSON_ARRAY_SIZE and JSON_OBJECT_SIZE macros
@@ -136,13 +135,13 @@ class ChatBox {
 public:
   // @param maxTokens
   // @param numMsgs
-  ChatBox(int maxTokens, int numMsgs);
+  ChatBox(uint16_t maxTokens, int numMsgs);
   ~ChatBox();
 
   bool init(const char* key, const char* model);
 
   // Getters
-  int maxTokens() const {
+  uint16_t maxTokens() const {
     return _maxTokens;
   }
 
@@ -198,7 +197,7 @@ public:
 
 
 private:
-  int _maxTokens;
+  uint16_t _maxTokens;
   int _maxMsgs;
   int _msgCount;
   int _MAX_MESSAGE_LENGTH;
